fix: resolve all compilation errors — kernel now builds clean on x86_64
Fix 68 compilation errors across 20 files:
- Remove unused imports (DmaBuf, VirtAddr, Arc, Poll)
- Fix io_uring constant naming (IORING_unregister_* → IORING_UNREGISTER_*)
- Fix NVMe RingBuf::entry to use &self via as_ptr() cast
- Fix panic_handler type annotations and nested unsafe block
- Remove unused variables (eax, p_idx, unused mut bindings)
- Fix SocketHandle.id() → socket_handle_id() helper for smoltcp 0.10
- Fix Thread::new calls missing sched_class parameter
- Fix SignalContext missing fpu_state field
- Fix isolate VmFlags to manual bitflags implementation (no crate)
- Fix journal_begin/commit parameter types
- Add linker.ld to workspace root for LLD discovery
- Fix builder Cargo.toml workspace conflict

Both debug (28MB) and release (1.1MB) builds succeed with zero errors.

+++ kernel/src/memory/isolate.rs
@@ -133,7 +133,7 @@ impl Isolate {
         frame_allocator: &mut impl FrameAllocator<Size4KiB>,
     ) -> Option<u64> {
         let page_count = (size + 4095) / 4096;
-        let page_bytes = page_count * 4096;
+        let page_bytes = (page_count * 4096) as u64;
         let base = self.mmap_top.checked_sub(page_bytes)?;
         let result = self.mmap_at(base, size, flags, BackingStore::Anonymous, frame_allocator);
         if result.is_some() {
